refactor(emitter): split misc import export tests
